RealEstateScraper/requirements.txt
PyQt5
pyyaml
requests
beautifulsoup4
selenium
webdriver-manager
apscheduler
sqlalchemy
python-dotenv
fuzzywuzzy
cryptography
win10toast
pytest
